# Implement PNG decode adaptor iterator