This incomplete PNG decoder presents as an Iterator reading u8 bytes directly from a Reader.

Approaching a png decoder as an Iterator requires significantly less Heap than
the traditional batch approach.

On Precursor, this enables PNG bytes to be intercepted at the network interface
and processed into the much smaller gam::Bitmap format without ever having to
store the original PNG file.