@nworbnhoj 2022-07-27
There is some benchmarking/tuning to be done at some stage with:

```
const IDAT_BUFFER_LENGTH: usize = 256;
const PNG_BUFFER_LENGTH: usize = 1024;
```
These 2 buffers consume heap space, but also probably impact speed.