@nworbnhoj 2022-07-27
The `miniz_oxide` `InflateState` struct is currently on the Stack.
```inflate_state: InflateState::new(DataFormat::Zlib),```
but it may be better on the Heap.
```inflate_state: InflateState::new_boxed(DataFormat::Zlib)```
So long as `IDAT_BUFFER_LENGTH` and `INFLATED_BUFFER_LENGTH` are relatively small then `InflateState` will also probably be small.