@nworbnhoj 2022-07-28
I just discovered that the `miniz_oxide` `InflateState` struct allocated an internal buffer of [u8; 32768] - so it is not going to work very well on the Stack or the Heap. This will need a little more work....