@bunnie 2022-07-29
I'm unable to fully test this because of #208, but I did check sizes with and without ditherpunk:

With
```
    PID   1:  628 k kernel
    PID   2: 1052 k gam
    PID   3:  784 k status
    PID   4:  796 k shellchat
    PID   5:  268 k ime-frontend
    PID   6:  164 k ime-plugin-shell
    PID   7:  880 k graphics-server
    PID   8:  208 k ticktimer-server
    PID   9:  128 k log-server
    PID  10:  228 k com
    PID  11:  176 k xous-names
    PID  12:  196 k keyboard
    PID  13:  184 k trng
    PID  14:  260 k llio
    PID  15:  172 k susres
    PID  16:  236 k codec
    PID  17:  176 k sha2
    PID  18:  200 k engine-25519
    PID  19:  220 k spinor
    PID  20:  832 k root-keys
    PID  21:  168 k jtag
    PID  22:  816 k net
    PID  23:  312 k dns
    PID  24: 1156 k pddb
    PID  25:  448 k modals
    PID  26:  268 k usb-device-xous
    PID  27: 1016 k vault
11972 k total
```

Without
```
    PID   1:  628 k kernel
    PID   2:  396 k gam
    PID   3:  780 k status
    PID   4:  692 k shellchat
    PID   5:  268 k ime-frontend
    PID   6:  164 k ime-plugin-shell
    PID   7:  488 k graphics-server
    PID   8:  208 k ticktimer-server
    PID   9:  128 k log-server
    PID  10:  228 k com
    PID  11:  176 k xous-names
    PID  12:  196 k keyboard
    PID  13:  184 k trng
    PID  14:  260 k llio
    PID  15:  172 k susres
    PID  16:  236 k codec
    PID  17:  176 k sha2
    PID  18:  200 k engine-25519
    PID  19:  220 k spinor
    PID  20:  828 k root-keys
    PID  21:  168 k jtag
    PID  22:  828 k net
    PID  23:  312 k dns
    PID  24: 1148 k pddb
    PID  25:  412 k modals
    PID  26:  272 k usb-device-xous
    PID  27: 1096 k vault
10864 k total
```

The good news is the bloat seems to be gone in `status` and `modals`! 

Most of the weight now sits in `gam` and `graphics-server` which is where you'd expect it to go. I'm not quite sure how you did it but that's awesome.