@bunnie 2022-08-04
Been wrestling with the network stack the past few days, and finally got it working well enough to do this:

![photo1659621300](https://user-images.githubusercontent.com/1168375/182864880-07a3c987-c27f-4c03-823a-15bb41e18082.jpeg)

slowly coming together...