@bunnie 2022-08-05
![photo1659687130](https://user-images.githubusercontent.com/1168375/183033184-1cb15c44-be57-403b-8bcc-a47a4f92acb0.jpeg)

Actually, I was on the wrong branch. This is what it looks like with your branch.

The performance is pretty slow -- takes maybe...3-5 minutes? to decode the PNG. I need to wrap a timer around it because my attention span isn't long enough to time it. But the memory usage seems pretty spot on: heap didn't go over 128k, which is really good.

It's the usual trade-off tho, isn't it? speed versus space. :-/