@bunnie 2022-08-05
Here's the runtime breakdown:

```
INFO:shellchat::cmds::net_cmd: fetching response.... (services\shellchat\src\cmds\net_cmd.rs:315)
INFO:shellchat::cmds::net_cmd: trying to read entire response... (services\shellchat\src\cmds\net_cmd.rs:317)
INFO:shellchat::cmds::net_cmd: "HTTP/1.1 200 OK\r\n" (services\shellchat\src\cmds\net_cmd.rs:347)
INFO:shellchat::cmds::net_cmd: "Date: Fri, 05 Aug 2022 07:57:46 GMT\r\n" (services\shellchat\src\cmds\net_cmd.rs:347)
...[snip]...
INFO:shellchat::cmds::net_cmd: "Content-Type: image/png\r\n" (services\shellchat\src\cmds\net_cmd.rs:347)
INFO:shellchat::cmds::net_cmd: attr: content-type,  image/png
 (services\shellchat\src\cmds\net_cmd.rs:351)
INFO:shellchat::cmds::net_cmd: found end of header after 10 lines. (services\shellchat\src\cmds\net_cmd.rs:336)
INFO:shellchat::cmds::net_cmd: heap size: 65536 (services\shellchat\src\cmds\net_cmd.rs:364)
INFO:gam::bitmap::decode_png: DecodePng ready: size(313,245) bit-depth=8 color_type=2 bytes_per_line=940 (services\gam\src\bitmap\decode_png.rs:126)
INFO:shellchat::cmds::net_cmd: png decoded in 10371ms (services\shellchat\src\cmds\net_cmd.rs:367)
INFO:shellchat::cmds::net_cmd: bitmap created (services\shellchat\src\cmds\net_cmd.rs:373)
INFO:gam::bitmap: rotate image (services\gam\src\bitmap.rs:183)
INFO:shellchat::cmds::net_cmd: converted to bitmap in 492825ms (services\shellchat\src\cmds\net_cmd.rs:376)
INFO:shellchat::cmds::net_cmd: heap size: 131072 (services\shellchat\src\cmds\net_cmd.rs:378)
INFO:shellchat::cmds::net_cmd: showing image (services\shellchat\src\cmds\net_cmd.rs:380)
INFO:gam::modal::image: bitmap height 312 : margin 8 (services\gam\src\modal\image.rs:35)
INFO:gam::modal::image: bitmap height 312 : margin 8 (services\gam\src\modal\image.rs:35)
INFO:gam::modal::image: drawing bitmap (services\gam\src\modal\image.rs:44)
INFO:graphics_server: ClipObject size: 4124 (services\graphics-server\src\lib.rs:223)
INFO:graphics_server: ClipObject size: 4124 (services\graphics-server\src\lib.rs:223)
INFO:graphics_server: ClipObject size: 4124 (services\graphics-server\src\lib.rs:223)
INFO:gam::modal::image: bitmap height 312 : margin 8 (services\gam\src\modal\image.rs:35)
INFO:gam::modal::image: drawing bitmap (services\gam\src\modal\image.rs:44)
...[snip]...
INFO:shellchat::cmds::net_cmd: done in 18482 ms (services\shellchat\src\cmds\net_cmd.rs:383)
```
The decode took 10 seconds; the conversion to bitmap took 492 seconds (about 8.2 minutes); the modal display was...less than 18 seconds, but it includes the time it took for me to notice it finished and dismiss the modal, so it could be much much faster than that.

Anyways, the long pole in the tent is the conversion to the bitmap, stuff behind this line of code:
```rust
let bm = gam::Bitmap::from_png(&mut png, Some(modal_size));
```

No big worries, I'm quite pleased with where this is all going tbh. Performance tuning is much easier imho than space tuning, I'm betting there's probably just an issue where data is being sent between processes or something and a little shuffling of abstractions will clear it all up. I'll have a look later, for now, just getting the net stack working to the point where I could decode the images reliably was surprisingly hard.