@bunnie 2022-08-07
I pulled in your latest `BufReader` changes, and the performance is much better now:

```
INFO:shellchat::cmds::net_cmd: png decoded in 621ms (services\shellchat\src\cmds\net_cmd.rs:365)
INFO:shellchat::cmds::net_cmd: bitmap created (services\shellchat\src\cmds\net_cmd.rs:371)
INFO:gam::bitmap: rotate image (services\gam\src\bitmap.rs:183)
INFO:shellchat::cmds::net_cmd: converted to bitmap in 25666ms (services\shellchat\src\cmds\net_cmd.rs:374)

```

- Decode time goes from 10371 ms -> 621 ms (94% improvement) 
- Render time goes from 492825ms -> 25666ms (95% improvement)

That's a nice victory lap for fixing BufReader. :tada: 

I'd say the decoder is fully usable now. The rendering needs a little more tweaking but a 19.2x speed improvement is already huge.